# Integrating Nested Prover Storage Management