--- a/adapters/mock-da/src/types/mod.rs
+++ b/adapters/mock-da/src/types/mod.rs
@@ -23,7 +23,7 @@ use crate::validity_condition::MockValidityCond;
     BorshDeserialize,
     BorshSerialize,
 )]
-pub struct MockHash([u8; 32]);
+pub struct MockHash(pub [u8; 32]);
 
 impl Debug for MockHash {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
@@ -59,7 +59,7 @@ impl std::hash::Hash for MockHash {
 impl BlockHashTrait for MockHash {}
 
 /// A mock block header used for testing.
-#[derive(Serialize, Deserialize, PartialEq, core::fmt::Debug, Clone)]
+#[derive(Serialize, Deserialize, PartialEq, Debug, Clone)]
 pub struct MockBlockHeader {
     /// The hash of the previous block.
     pub prev_hash: MockHash,
@@ -71,14 +71,24 @@ pub struct MockBlockHeader {
     pub time: Time,
 }
 
-impl Default for MockBlockHeader {
-    fn default() -> Self {
-        Self {
-            prev_hash: MockHash([0u8; 32]),
-            hash: MockHash([1u8; 32]),
-            height: 0,
+impl MockBlockHeader {
+    /// Generates [`MockBlockHeader`] with given height, where hashes are derived from height
+    /// Can be used in tests, where header of following blocks will be consistent
+    pub fn from_height(height: u64) -> MockBlockHeader {
+        let prev_hash = u64_to_bytes(height);
+        let hash = u64_to_bytes(height + 1);
+        MockBlockHeader {
+            prev_hash: MockHash(prev_hash),
+            hash: MockHash(hash),
+            height,
             time: Time::now(),
         }
+    }
+}
+
+impl Default for MockBlockHeader {
+    fn default() -> Self {
+        MockBlockHeader::from_height(0)
     }
 }
 
@@ -172,7 +182,7 @@ impl MockBlob {
 }
 
 /// A mock block type used for testing.
-#[derive(Serialize, Deserialize, PartialEq, core::fmt::Debug, Clone)]
+#[derive(Serialize, Deserialize, Default, PartialEq, Debug, Clone)]
 pub struct MockBlock {
     /// The header of this block.
     pub header: MockBlockHeader,
@@ -180,21 +190,6 @@ pub struct MockBlock {
     pub validity_cond: MockValidityCond,
     /// Blobs
     pub blobs: Vec<MockBlob>,
-}
-
-impl Default for MockBlock {
-    fn default() -> Self {
-        Self {
-            header: MockBlockHeader {
-                prev_hash: [0; 32].into(),
-                hash: [1; 32].into(),
-                height: 0,
-                time: Default::default(),
-            },
-            validity_cond: Default::default(),
-            blobs: Default::default(),
-        }
-    }
 }
 
 impl SlotData for MockBlock {
@@ -212,4 +207,21 @@ impl SlotData for MockBlock {
     fn validity_condition(&self) -> MockValidityCond {
         self.validity_cond
     }
+}
+
+impl MockBlock {
+    /// Creates empty block, which is following of the current
+    pub fn next_mock(&self) -> MockBlock {
+        let mut next_block = MockBlock::default();
+        let h = self.header.height + 1;
+        next_block.header = MockBlockHeader::from_height(h);
+        next_block
+    }
+}
+
+fn u64_to_bytes(value: u64) -> [u8; 32] {
+    let value = value.to_be_bytes();
+    let mut result = [0u8; 32];
+    result[..value.len()].copy_from_slice(&value);
+    result
 }